+++ src/systems/a320_systems/src/hydraulic.rs
@@ -5,7 +5,7 @@ use uom::si::{
 };
 
 use systems::overhead::{
-    AutoOffFaultPushButton, AutoOnFaultPushButton, MomentaryPushButton, OnOffFaultPushButton,
+    AutoOffFaultPushButton, AutoOnFaultPushButton, MomentaryOnPushButton, MomentaryPushButton,
 };
 use systems::{
     hydraulic::{
@@ -1329,7 +1329,8 @@ pub(super) struct A320HydraulicOverheadPanel {
     ptu_push_button: AutoOffFaultPushButton,
     rat_push_button: MomentaryPushButton,
     yellow_epump_push_button: AutoOnFaultPushButton,
-    blue_epump_override_push_button: OnOffFaultPushButton,
+    blue_epump_override_push_button: MomentaryOnPushButton,
+    blue_pump_override_relay_powered_by: ElectricalBusType,
 }
 impl A320HydraulicOverheadPanel {
     pub(super) fn new() -> A320HydraulicOverheadPanel {
@@ -1340,7 +1341,8 @@ impl A320HydraulicOverheadPanel {
             ptu_push_button: AutoOffFaultPushButton::new_auto("HYD_PTU"),
             rat_push_button: MomentaryPushButton::new("HYD_RAT_MAN_ON"),
             yellow_epump_push_button: AutoOnFaultPushButton::new_auto("HYD_EPUMPY"),
-            blue_epump_override_push_button: OnOffFaultPushButton::new_off("HYD_EPUMPY_OVRD"),
+            blue_epump_override_push_button: MomentaryOnPushButton::new("HYD_EPUMPY_OVRD"),
+            blue_pump_override_relay_powered_by: A320Hydraulic::BLUE_ELEC_PUMP_CONTROL_POWER_BUS,
         }
     }
 
@@ -1402,6 +1404,20 @@ impl SimulationElement for A320HydraulicOverheadPanel {
         self.blue_epump_override_push_button.accept(visitor);
 
         visitor.visit(self);
+    }
+
+    fn receive_power(&mut self, supplied_power: &SuppliedPower) {
+        let blue_pump_override_relay_is_powered = supplied_power
+            .potential_of(&self.blue_pump_override_relay_powered_by)
+            .is_powered();
+
+        // Relay truth table is actually a Xor, as output is 1 if ON and not(PRESSED) or if PRESSED and not(ON)
+        let blue_pump_override_relay_output = (self.blue_epump_override_push_button.is_on()
+            ^ self.blue_epump_override_push_button.is_pressed())
+            && blue_pump_override_relay_is_powered;
+
+        self.blue_epump_override_push_button
+            .set_on(blue_pump_override_relay_output);
     }
 }
 
@@ -1829,6 +1845,11 @@ mod tests {
                     .read_bool("OVHD_HYD_EPUMPB_PB_HAS_FAULT")
             }
 
+            fn blue_epump_override_is_on(&mut self) -> bool {
+                self.simulation_test_bed
+                    .read_bool("OVHD_HYD_EPUMPY_OVRD_IS_ON")
+            }
+
             fn get_brake_left_yellow_pressure(&mut self) -> Pressure {
                 Pressure::new::<psi>(
                     self.simulation_test_bed
@@ -2066,9 +2087,9 @@ mod tests {
                 self
             }
 
-            fn set_blue_e_pump_ovrd(mut self, is_on: bool) -> Self {
+            fn set_blue_e_pump_ovrd_pressed(mut self, is_pressed: bool) -> Self {
                 self.simulation_test_bed
-                    .write_bool("OVHD_HYD_EPUMPY_OVRD_PB_IS_ON", is_on);
+                    .write_bool("OVHD_HYD_EPUMPY_OVRD_IS_PRESSED", is_pressed);
                 self
             }
 
@@ -2124,9 +2145,13 @@ mod tests {
                 self
             }
 
+            fn dc_ess_active(mut self) -> Self {
+                self.aircraft.set_dc_ess_is_powered(true);
+                self
+            }
+
             fn set_cold_dark_inputs(self) -> Self {
-                self.set_blue_e_pump_ovrd(false)
-                    .set_eng1_fire_button(false)
+                self.set_eng1_fire_button(false)
                     .set_eng2_fire_button(false)
                     .set_blue_e_pump(true)
                     .set_yellow_e_pump(true)
@@ -2207,6 +2232,14 @@ mod tests {
                     .set_left_brake(Ratio::new::<percent>(0.))
                     .set_right_brake(Ratio::new::<percent>(0.))
                     .run_waiting_for(Duration::from_secs(1));
+
+                self
+            }
+
+            fn press_blue_epump_override_button_once(mut self) -> Self {
+                self = self.set_blue_e_pump_ovrd_pressed(true).run_one_tick();
+
+                self = self.set_blue_e_pump_ovrd_pressed(false).run_one_tick();
 
                 self
             }
@@ -2667,7 +2700,8 @@ mod tests {
             // Blue epump should have no fault
             assert!(!test_bed.is_blue_epump_press_low_fault());
 
-            test_bed = test_bed.set_blue_e_pump_ovrd(true).run_one_tick();
+            test_bed = test_bed.press_blue_epump_override_button_once();
+            assert!(test_bed.blue_epump_override_is_on());
 
             // As we use override, this bypasses eng off fault inhibit so we have a fault
             assert!(test_bed.is_blue_epump_press_low_fault());
@@ -2947,7 +2981,8 @@ mod tests {
             assert!(!test_bed.is_blue_epump_press_low());
 
             // Starting epump
-            test_bed = test_bed.set_blue_e_pump_ovrd(true).run_one_tick();
+            test_bed = test_bed.press_blue_epump_override_button_once();
+            assert!(test_bed.blue_epump_override_is_on());
 
             // Pump commanded on but pressure couldn't rise enough: we are in fault low
             assert!(test_bed.is_blue_epump_press_low());
@@ -2961,10 +2996,39 @@ mod tests {
             assert!(!test_bed.is_blue_epump_press_low());
 
             // Stoping epump, no fault expected
-            test_bed = test_bed
-                .set_blue_e_pump_ovrd(false)
-                .run_waiting_for(Duration::from_secs(1));
+            test_bed = test_bed.press_blue_epump_override_button_once();
+            assert!(!test_bed.blue_epump_override_is_on());
+
+            test_bed = test_bed.run_waiting_for(Duration::from_secs(1));
             assert!(!test_bed.is_blue_epump_press_low());
+        }
+
+        #[test]
+        fn blue_epump_override_switches_to_off_when_losing_relay_power_and_stays_off() {
+            let mut test_bed = test_bed_with()
+                .engines_off()
+                .on_the_ground()
+                .set_cold_dark_inputs()
+                .run_one_tick();
+
+            // Starting epump
+            test_bed = test_bed
+                .press_blue_epump_override_button_once()
+                .run_waiting_for(Duration::from_secs(10));
+            assert!(test_bed.blue_epump_override_is_on());
+            assert!(test_bed.is_blue_pressurised());
+
+            // Killing the bus corresponding to the latching relay of blue pump override push button
+            // It should set the override state back to off without touching the push button
+            test_bed = test_bed.dc_ess_lost().run_one_tick();
+            assert!(!test_bed.blue_epump_override_is_on());
+
+            // Stays off even powered back
+            test_bed = test_bed.dc_ess_active().run_one_tick();
+            assert!(!test_bed.blue_epump_override_is_on());
+
+            test_bed = test_bed.run_waiting_for(Duration::from_secs(10));
+            assert!(!test_bed.is_blue_pressurised());
         }
 
         #[test]
@@ -3287,17 +3351,20 @@ mod tests {
                 .run_one_tick();
 
             // Starting blue_epump wait for pressure rise to make sure system is primed
-            test_bed = test_bed
-                .set_blue_e_pump_ovrd(true)
-                .run_waiting_for(Duration::from_secs(20));
+            test_bed = test_bed.press_blue_epump_override_button_once();
+            assert!(test_bed.blue_epump_override_is_on());
+
+            test_bed = test_bed.run_waiting_for(Duration::from_secs(20));
             assert!(test_bed.is_blue_pressurised());
             assert!(test_bed.blue_pressure() < Pressure::new::<psi>(3500.));
             assert!(test_bed.blue_pressure() > Pressure::new::<psi>(2500.));
 
             // Shutdown and wait for pressure stabilisation
-            test_bed = test_bed
-                .set_blue_e_pump_ovrd(false)
-                .run_waiting_for(Duration::from_secs(50));
+            test_bed = test_bed.press_blue_epump_override_button_once();
+            assert!(!test_bed.blue_epump_override_is_on());
+
+            test_bed = test_bed.run_waiting_for(Duration::from_secs(50));
+
             assert!(!test_bed.is_blue_pressurised());
             assert!(test_bed.blue_pressure() < Pressure::new::<psi>(50.));
             assert!(test_bed.blue_pressure() > Pressure::new::<psi>(-50.));
@@ -3839,8 +3906,7 @@ mod tests {
         #[test]
         fn controller_blue_epump_activates_when_no_weight_on_wheels() {
             let engine_off_oil_pressure = Pressure::new::<psi>(10.);
-            let mut overhead_panel = A320HydraulicOverheadPanel::new();
-            overhead_panel.blue_epump_override_push_button.push_off();
+            let overhead_panel = A320HydraulicOverheadPanel::new();
 
             let mut blue_epump_controller = A320BlueElectricPumpController::new(
                 A320Hydraulic::BLUE_ELEC_PUMP_CONTROL_POWER_BUS,
@@ -3882,8 +3948,7 @@ mod tests {
         fn controller_blue_epump_split_engine_states() {
             let engine_on_oil_pressure = Pressure::new::<psi>(30.);
             let engine_off_oil_pressure = Pressure::new::<psi>(10.);
-            let mut overhead_panel = A320HydraulicOverheadPanel::new();
-            overhead_panel.blue_epump_override_push_button.push_off();
+            let overhead_panel = A320HydraulicOverheadPanel::new();
 
             let mut blue_epump_controller = A320BlueElectricPumpController::new(
                 A320Hydraulic::BLUE_ELEC_PUMP_CONTROL_POWER_BUS,
@@ -3934,8 +3999,7 @@ mod tests {
         fn controller_blue_epump_on_off_engines() {
             let engine_on_oil_pressure = Pressure::new::<psi>(30.);
             let engine_off_oil_pressure = Pressure::new::<psi>(10.);
-            let mut overhead_panel = A320HydraulicOverheadPanel::new();
-            overhead_panel.blue_epump_override_push_button.push_off();
+            let overhead_panel = A320HydraulicOverheadPanel::new();
 
             let mut blue_epump_controller = A320BlueElectricPumpController::new(
                 A320Hydraulic::BLUE_ELEC_PUMP_CONTROL_POWER_BUS,
@@ -3985,7 +4049,7 @@ mod tests {
 
             let eng1_above_idle = false;
             let eng2_above_idle = false;
-            overhead_panel.blue_epump_override_push_button.push_on();
+            overhead_panel.blue_epump_override_push_button.set_on(true);
             blue_epump_controller.update(
                 &overhead_panel,
                 true,
@@ -3998,7 +4062,7 @@ mod tests {
 
             let eng1_above_idle = false;
             let eng2_above_idle = false;
-            overhead_panel.blue_epump_override_push_button.push_off();
+            overhead_panel.blue_epump_override_push_button.set_on(false);
             blue_epump_controller.update(
                 &overhead_panel,
                 false,
@@ -4021,7 +4085,7 @@ mod tests {
 
             let eng1_above_idle = false;
             let eng2_above_idle = false;
-            overhead_panel.blue_epump_override_push_button.push_on();
+            overhead_panel.blue_epump_override_push_button.set_on(true);
 
             blue_epump_controller.receive_power(&test_supplied_power(
                 ElectricalBusType::DirectCurrentEssential,