+++ .github/CHANGELOG.md
@@ -78,6 +78,7 @@
 1. [MCDU] Fixed sometimes bad error message on crz fl entry - @MisterChocker (Leon)
 1. [ECP] Corrected EMER CANC ECAM button emissive behavior - @ImenesFBW (Imenes)
 1. [ECAM] Change flight control page to use hydraulic simulation vars - @lukecologne (luke)
+1. [HYD] Blue electric pump override button changed to momentary latched - @Crocket63 (crocket)
 
 ## 0.6.0
 1. [CDU] Added WIND page - @tyler58546 (tyler58546)